# api-docs/swagger.json
{
  "info": {
    "title": "Staffing API",
    "description": "Staffing API Sandbox",
    "version": "1.0.0"
  },
  "schemes": [
    "https"
  ],
  "basePath": "/api",
  "swagger": "2.0",
  "paths": {
...
    "/proposals": {
      "get": {
        "tags": [
          "Proposal"
        ],
        "summary": "List all proposals",
        "description": "List all proposals.",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful Operation.",
            "schema": {
              "$ref": "#/definitions/Proposal"
            }
          },
          "401": {
            "description": "unauthorized"
          },
          "500": {
            "description": "internal error"
          }
        }
      },
      "post": {
        "tags": [
          "Proposal"
        ],
        "summary": "Create proposal",
        "description": "Create proposal with form fields",
        "produces": [
          "application/json"
        ],
        "parameters": {
          "name": {
            "type": "string"
          },
          "subhead": {
            "type": "string"
          },
          "overview": {
            "type": "string"
          },
          "overview_inactive": {
            "type": "boolean"
          },
          "scope_of_work": {
            "type": "string"
          },
          "scope_of_work_inactive": {
            "type": "boolean"
          },
          "inserts": {
            "type": "array",
            "items": {
              "type": "object",
              "schema": {
                "$ref": "#/definitions/Insert"
              }
            }
          },
          "inserts_inactive": {
            "type": "boolean"
          },
          "breakdown": {
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Breakdown"
            }
          },
          "breakdown_inactive": {
            "type": "boolean"
          },
          "timeline": {
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Breakdown"
            }
          },
          "timeline_inactive": {
            "type": "boolean"
          },
          "steps": {
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Steps"
            }
          },
          "steps_inactive": {
            "type": "boolean"
          },
          "candidate_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "required": true
            }
          },
          "candidate_ids_inactive": {
            "type": "boolean"
          },
          "portfolio_ids": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "portfolio_ids_inactive": {
            "type": "boolean"
          }
        },
        "responses": {
          "200": {
            "description": "Successful Operation.",
            "schema": {
              "$ref": "#/definitions/Proposal"
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "put": {
        "tags": [
          "Proposal"
        ],
        "summary": "Create proposal",
        "description": "Create proposal with form fields",
        "produces": [
          "application/json"
        ],
        "parameters": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "subhead": {
            "type": "string"
          },
          "overview": {
            "type": "string"
          },
          "overview_inactive": {
            "type": "boolean"
          },
          "scope_of_work": {
            "type": "string"
          },
          "scope_of_work_inactive": {
            "type": "boolean"
          },
          "inserts": {
            "type": "array",
            "items": {
              "type": "object",
              "schema": {
                "$ref": "#/definitions/Insert"
              }
            }
          },
          "inserts_inactive": {
            "type": "boolean"
          },
          "breakdown": {
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Breakdown"
            }
          },
          "breakdown_inactive": {
            "type": "boolean"
          },
          "timeline": {
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Breakdown"
            }
          },
          "timeline_inactive": {
            "type": "boolean"
          },
          "steps": {
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Steps"
            }
          },
          "steps_inactive": {
            "type": "boolean"
          },
          "candidate_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "required": true
            }
          },
          "candidate_ids_inactive": {
            "type": "boolean"
          },
          "portfolio_ids": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "portfolio_ids_inactive": {
            "type": "boolean"
          }
        },
        "responses": {
          "200": {
            "description": "Successful Operation.",
            "schema": {
              "$ref": "#/definitions/Proposal"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal error"
          }
        }
      }
    },
    "/proposals/:id": {
      "get": {
        "tags": [
          "Proposal"
        ],
        "summary": "Get proposal by id",
        "description": "Get proposal by id",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful Operation.",
            "schema": {
              "$ref": "#/definitions/Proposal"
            }
          },
          "401": {
            "description": "unauthorized"
          },
          "500": {
            "description": "internal error"
          }
        }
      },
      "delete": {
        "tags": [
          "Proposal"
        ],
        "summary": "Delete proposal by id",
        "description": "Delete proposal by id",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful Operation."
          },
          "401": {
            "description": "unauthorized"
          },
          "500": {
            "description": "internal error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Breakdown": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "summary": {
          "type": "string"
        },
        "tables": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "row": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "Insert": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "paragraph": {
          "type": "string"
        }
      }
    },
    "Steps": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "summary": {
          "type": "string"
        },
        "list": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "header": "string",
              "paragraph": "string"
            }
          }
        }
      }
    },
    "Proposal": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "subhead": {
          "type": "string"
        },
        "overview": {
          "type": "string"
        },
        "overview_inactive": {
          "type": "boolean"
        },
        "scope_of_work": {
          "type": "string"
        },
        "scope_of_work_inactive": {
          "type": "boolean"
        },
        "inserts": {
          "type": "array",
          "items": {
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Insert"
            }
          }
        },
        "inserts_inactive": {
          "type": "boolean"
        },
        "breakdown": {
          "type": "object",
          "schema": {
            "$ref": "#/definitions/Breakdown"
          }
        },
        "breakdown_inactive": {
          "type": "boolean"
        },
        "timeline": {
          "type": "object",
          "schema": {
            "$ref": "#/definitions/Breakdown"
          }
        },
        "timeline_inactive": {
          "type": "boolean"
        },
        "steps": {
          "type": "object",
          "schema": {
            "$ref": "#/definitions/Steps"
          }
        },
        "steps_inactive": {
          "type": "boolean"
        },
        "candidate_ids": {
          "type": "array",
          "items": {
            "type": "string",
            "required": true
          }
        },
        "candidate_ids_inactive": {
          "type": "boolean"
        },
        "portfolio_ids": {
          "type": "array",
          "items": {
            "type": "number"
          }
        },
        "portfolio_ids_inactive": {
          "type": "boolean"
        }
      }
    },
...
  },
...
  "securityDefinitions": {}
}